+++ apps/restconf/restconf_methods.c
@@ -333,7 +333,7 @@ api_data_write(clicon_handle h,
 	}
 	break;
     case YANG_DATA_JSON:
-	if ((ret = clixon_json_parse_string(data, yb, yspec, &xdata0, &xerr)) < 0){
+	if ((ret = clixon_json_parse_string(data, 1, yb, yspec, &xdata0, &xerr)) < 0){
 	    if (netconf_malformed_message_xml(&xerr, clicon_err_reason) < 0)
 		goto done;
 	    if (api_return_err0(h, req, xerr, pretty, media_out, 0) < 0)