+++ util/clixon_util_socket.c
@@ -149,7 +149,7 @@ main(int    argc,
     }
     /* 2. Parse data (xml/json) */
     if (jsonin){
-	if ((ret = clixon_json_parse_file(fp, YB_NONE, NULL, &xt, &xerr)) < 0)
+	if ((ret = clixon_json_parse_file(fp, 0, YB_NONE, NULL, &xt, &xerr)) < 0)
 	    goto done;
 	if (ret == 0){
 	    fprintf(stderr, "Invalid JSON\n");